Исправления с учетом комментариев ревьювера
