+++ MovieQuiz/Services/QuestionFactory.swift
@@ -11,6 +11,9 @@ final class QuestionFactory: QuestionFactoryProtocol {
     private let moviesLoader: MoviesLoading
     weak var delegate: QuestionFactoryDelegate?
     
+    let movieQuizViewController = MovieQuizViewController()
+    
+    
     init(moviesLoader: MoviesLoading, delegate: QuestionFactoryDelegate?) {
         self.moviesLoader = moviesLoader
         self.delegate = delegate
@@ -25,8 +28,18 @@ final class QuestionFactory: QuestionFactoryProtocol {
                 guard let self = self else { return }
                 switch result {
                 case .success(let mostPopularMovies):
-                    self.movies = mostPopularMovies.items
-                    self.delegate?.didLoadDataFromServer()
+                    // Если массив фильмов оказался пустым
+                    if mostPopularMovies.items.isEmpty {
+                        self.delegate?.didFailToLoadData(with: "Ошибка загрузки перечня фильмов" as! any Error)
+                    // Если есть сообщение об ошибке
+                    } else if mostPopularMovies.errorMessage != "" {
+                        self.delegate?.didFailToLoadData(with: "\(mostPopularMovies.errorMessage)" as! any Error)
+                    } else {
+                    // Если все ошибок нет
+                        self.movies = mostPopularMovies.items
+                        self.delegate?.didLoadDataFromServer()
+                    }
+                // Если есть ошибки с кодом
                 case .failure(let error):
                     self.delegate?.didFailToLoadData(with: error)
                 }
@@ -49,6 +62,10 @@ final class QuestionFactory: QuestionFactoryProtocol {
                 imageData = try Data(contentsOf: movie.resizedImage)
             } catch {
                 print("Failed to load image")
+                // Передача информации об ошибке загрузки в основной поток
+                DispatchQueue.main.async{
+                    self.delegate?.didFailToLoadData(with: error)
+                }
             }
             
             // Определяем рейтинг, рейтинг разный для каждого вопроса
@@ -68,47 +85,3 @@ final class QuestionFactory: QuestionFactoryProtocol {
         }
     }
 }
-
-
-// Массив mock`ов
-//private let questions: [QuizQuestion] = [
-//        QuizQuestion(
-//            image: "The Godfather",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "The Dark Knight",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "Kill Bill",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "The Avengers",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "Deadpool",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "The Green Knight",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: true),
-//        QuizQuestion(
-//            image: "Old",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: false),
-//        QuizQuestion(
-//            image: "The Ice Age Adventures of Buck Wild",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: false),
-//        QuizQuestion(
-//            image: "Tesla",
-//            text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: false),
-//        QuizQuestion(
-//            image: "Vivarium", text: "Рейтинг этого фильма больше чем 6?",
-//            correctAnswer: false)
-//    ]
